Refactor comment and separate code for querying the Edamam Recipe API.
